implement TradingEquipment component
