--- practice-app/frontend/mercatus/src/TradingEquipment/comment.css
@@ -1,10 +1,0 @@
-.comment-container {
-    border-width: 2px;
-    border-style: solid;
-
-    margin: 5px;
-}
-
-.author {
-    border-bottom: 1px solid red;
-}
