+++ practice-app/frontend/mercatus/src/components/TradingEquipment/tr-eq.css
@@ -1,0 +1,6 @@
+.symbol-header {
+    text-transform: uppercase;
+    font-size: 20px;
+    font-weight: 600;
+    margin: 20px;
+}
